Add Action metadata file to the root directory
As required by GitHub Actions for Action repositories, I've placed the
action metadata file, action.yaml, in the root directory of the
repository.

The currently defined arguments are the minimally required arguments to
make the action run. There are no additional settings to configure
things like embed color, webhook author, or webhook avatar yet.

As the docker application will get the inputs as positional arguments,
I'll add a feature that parses this file and adds each input to the
argparse argument parser automatically. That way, the arguments and
their descriptions don't have to be maintained in two locations. It does
mean that the `args` list has to be specified in the same order as the
one the input arguments are defined in.