Add dataclasses to assist parsing action inputs
I've added dataclasses that will assist in parsing the input the action
will get from the command line entrypoint.

The dataclasses inherit from a base class that implements an alternative
constructor, `from_arguments`, that will create an instance by consuming
the appropriate arguments from a dictionary with CLI arguments.

The `from_arguments` constructor will attempt to convert the arguments
to the appropriate types, which means we'll fail early if arguments were
incorrectly specified. (A future version may receive stricter validation
rules.)

One feature of the alternative constructor is that it returns `None` if
none of the arguments for an *optional* dataset were given instead of
creating an instance where every attribute is `None`. This allows us to
easily check for whether or not pull request information was included.