Parse CLI arguments from Action metadata
I've added an argparse CLI parser with dynamically generated arguments
based on the `action.yaml` metadata file. This means that all inputs
that we get automatically receive a CLI argument in our Python package,
with a descriptive help message hint.

There's one small complication: If this application is run directly
within a git repository, the `action.yaml` file is located in the root
of the repository (as required by GitHub Actions). However, when we
package the application (e.g., docker of pypi), the action.yaml file
will be located within the package directory.

To solve that, the application will check both possible locations, in
the order of "package" -> "parent directory". If the file is not found
at either location, a new exception, MissingActionFile, is raised.