Create webhook payload with helper functions
I've created two helper functions, `get_payload_pull_request` and
`get_payload` to construct payloads for the webhook endpoint using our
TypedDict classes.

If we want to communicate the status of a check belonging to a PR, we'll
use the pull request helper; if not, we'll use the regular helper
function.