Send the payload to the webhook using requests
I've added a function that actually sends the generated payload to the
webhook endpoint using `requests`.

If the request results in an error status, this function will print an
error message with (hopefully) some information as a GitHub Actions
error message.

If the request results in an "ok" status, it prints a "success" message
to stdout.

At the end of the function, a `bool` is returned to indicate success or
failure.