Estimate the characters available for source field
The source field in a PR embed gets truncated down to a certain number
of characters to hopefully make it fit on one line. However, the
available space depends on the length of the other two fields in its
line.

I'm now trying to estimate the number of characters left for the source
field by allowing it a character budget based on the other fields. This
is only an estimation, as the actual display width depends on the
specific characters used (i.e. Discord uses a non-fixed width font).