Use defaults for args available in environment
A number of arguments that the action needs are available using the
`github` context provided in a workflow. While I still want to make keep
the embed fully configurable, I've now set default values that make
sense in most situations.

One caveat is that if the embed is sent from a `workflow_run` event
(e.g., to allow for accessing secrets), the `github` context will be
that of the `workflow_run` trigger; not the workflow that triggered the
workflow_run-triggered workflow.