Add deprecation warning for `pull_request_payload`
Unfortunately, passing JSON serialized strings to the Python package as
command line arguments parsed by argparse led to some issues with escape
sequences getting mangled.

While I've introduced a fix for the issue, it's quite fragile and I'm
not confident enough in the fix that I want to continue supporting it in
the first major release.

That's why I've added a deprecation warning to communicate that it will
be removed in version 1.0.0.